Fixed #230: Unknown Inverter Type SBS3.7-10 Fixed #261: Unknown Inverter Type STP4.0-3AV-40 Fixed #269: SMA STP6.0-3AV-40 no readout with SBFspot? Fixed #279: Unknown inverter type SMA STP6.0-3AV-40 Fixed #181: SBFspotUploadDaemon not uploading

+++ SBFspot/SBFspot.h
@@ -449,7 +449,7 @@ const char *getEventType(unsigned short eventflags);
 int getInverterData(InverterData *inverters[], enum getInverterDataType type);
 int getInverterIndexByAddress(InverterData *inverters[], unsigned char bt_addr[6]);
 int getInverterIndexBySerial(InverterData *inverters[], unsigned short SUSyID, uint32_t Serial);
-int getInverterIndexBySerial(InverterData *inverters[], unsigned long Serial);
+int getInverterIndexBySerial(InverterData *inverters[], uint32_t Serial);
 E_SBFSPOT getPacket(unsigned char senderaddr[6], int wait4Command);
 //int get_tzOffset(void);
 void HexDump(unsigned char *buf, int count, int radix);
